ArgumentMap -> ReducerMap
